+++ Sources/MapboxNavigation/InstructionsCardViewController.swift
@@ -491,7 +491,7 @@ extension InstructionsCardViewController: NavigationMapInteractionObserver {
         stopPreview()
     }
     
-    public func navigationViewController(updateTo styleURI: StyleURI?) {
+    public func navigationViewController(willUpdateTo styleURI: StyleURI?) {
         spriteRepository.updateStyle(styleURI: styleURI) { [weak self] in
             guard let self = self else { return }
             self.updateLegacyIcons()